requirements.txt
gymnasium>=0.29.0
stable-baselines3>=2.0.0
torch>=1.13.0
pygame>=2.1.0
numpy>=1.21.0
matplotlib>=3.5.0
tensorboard>=2.10.0
Pillow>=9.0.0
scipy>=1.9.0